+++ mock/MockProvider.js
@@ -1,127 +1,58 @@
-/* eslint-disable camelcase */
+/* eslint-disable camelcase,no-unused-vars */
+
 const EventEmitter = require('events');
 const lodash = require('lodash');
-const { toHex, padHex, randomPick, randomHex, HexStruct } = require('./util');
+const { toHex, padHex, randomHex, HexStruct } = require('./util');
+const { Contract, ChecksumAddress } = require('../src');
 
-const addressStruct = new HexStruct('0xa0', { address: 4 }, 40);
-const blockHashStruct = new HexStruct('0xb0', { epochNumber: 6, blockIndex: 2 }, 64);
-const txHashStruct = new HexStruct('0xf0', { epochNumber: 6, blockIndex: 2, transactionIndex: 4 }, 64);
+const contract = new Contract({
+  abi: [
+    {
+      anonymous: false,
+      type: 'event',
+      name: 'Transfer',
+      inputs: [
+        { type: 'address', name: 'from', indexed: true },
+        { type: 'address', name: 'to', indexed: true },
+        { type: 'uint256', name: 'value' },
+      ],
+    },
+  ],
+});
 
-/**
- * @param self {MockProvider}
- * @param epochNumber {number}
- * @return {string[]}
- */
-function mockBlockHashArray(self, epochNumber) {
-  const blockHashArray = lodash.range(self.epochBlockCount).map(
-    i => blockHashStruct.encode({ epochNumber, blockIndex: i }),
-  );
-  return blockHashArray.reverse(); // pivot to last one
-}
+const accountAddressStruct = new HexStruct('0x10', { address: 6 }, 40);
+const contractAddressStruct = new HexStruct('0x80', { epochNumber: 6, blockIndex: 6 }, 40);
+const tokenAddressStruct = new HexStruct('0x80', { index: 6 }, 40);
+const blockHashStruct = new HexStruct('0xb0', { epochNumber: 6, blockIndex: 6 }, 64);
+const txHashStruct = new HexStruct('0xf0', { epochNumber: 6, blockIndex: 6, transactionIndex: 6 }, 64);
 
-function mockBlockByHash(self, blockHash) {
-  const { epochNumber, blockIndex } = blockHashStruct.decode(blockHash);
-
-  const blockCount = epochNumber * self.epochBlockCount + blockIndex;
-  const minerIndex = blockCount % self.addressCount;
-
-  const parentHash = blockIndex === 0
-    ? blockHashStruct.encode({ epochNumber: epochNumber ? epochNumber - 1 : 0, blockIndex })
-    : blockHashStruct.encode({ epochNumber, blockIndex: blockIndex - 1 });
-
-  const refereeHashes = [blockHashStruct.encode({ epochNumber, blockIndex })];
-  const miner = addressStruct.encode({ address: minerIndex });
-  const timestamp = self.startTimestamp + (blockCount * self.blockDelta); // in secords
-
-  return {
-    epochNumber: toHex(epochNumber),
-    parentHash,
-    refereeHashes,
-    miner,
-    timestamp: toHex(timestamp),
-  };
-}
-
-function mockTxHashArray(self, blockHash) {
-  const { epochNumber, blockIndex } = blockHashStruct.decode(blockHash);
-
-  return lodash.range(self.blockTxCount).map(
-    i => txHashStruct.encode({ epochNumber, blockIndex, transactionIndex: i }),
-  );
-}
-
-function mockTxByHash(self, transactionHash) {
-  const { epochNumber, blockIndex, transactionIndex } = txHashStruct.decode(transactionHash);
-
-  const blockCount = epochNumber * self.epochBlockCount + blockIndex;
-  const txCount = blockCount * self.blockTxCount + transactionIndex;
-  const fromIndex = txCount % self.addressCount;
-  const toIndex = (txCount + 1) % self.addressCount;
-
-  const blockHash = blockHashStruct.encode({ epochNumber, blockIndex });
-  const nonce = Math.floor(txCount / self.addressCount);
-  const from = addressStruct.encode({ address: fromIndex });
-  const to = addressStruct.encode({ address: toIndex });
-
-  return {
-    epochNumber,
-    blockHash,
-    transactionIndex: toHex(transactionIndex),
-    from,
-    nonce: toHex(nonce),
-    to,
-    status: self.stable ? '0x0' : randomPick(null, '0x0', '0x1'),
-  };
-}
-
-function mockLogsByEpochNumber(self, epochNumber) {
-  const [blockHash] = mockBlockHashArray(self, epochNumber);
-  if (!blockHash) {
-    return [];
-  }
-
-  const [transactionHash] = mockTxHashArray(self, blockHash);
-  if (!transactionHash) {
-    return [];
-  }
-
-  const address = addressStruct.encode({ address: epochNumber });
-  return lodash.range(self.epochTxCount).map(index => ({
-    address,
-    epochNumber: toHex(epochNumber),
-    blockHash,
-    logIndex: randomHex(1),
-    transactionHash,
-    transactionIndex: toHex(0),
-    transactionLogIndex: toHex(index),
-    topics: [
-      '0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef',
-      padHex(epochNumber, 64),
-      padHex(epochNumber + 1, 64),
-    ],
-    data: randomHex(64),
-  }));
-}
-
-// ----------------------------------------------------------------------------
 class MockProvider extends EventEmitter {
   constructor({
     startTimestamp = Math.floor(Date.now() / 1000) - 2 * 30 * 24 * 3600,
-    blockDelta = 1,
+    epochNumber = Number.MAX_SAFE_INTEGER,
+    blockDelta = 1, // in secords
     addressCount = 10,
     epochBlockCount = 5,
     blockTxCount = 2,
     epochTxCount = 2,
-    stable = true,
+    chainId = 1029,
+    netName = 'CFX',
   } = {}) {
     super();
+
     this.startTimestamp = startTimestamp;
+    this.epochNumber = epochNumber;
     this.blockDelta = blockDelta;
-    this.addressCount = addressCount;
     this.epochBlockCount = epochBlockCount;
     this.blockTxCount = blockTxCount;
     this.epochTxCount = epochTxCount;
-    this.stable = stable;
+    this.chainId = chainId;
+    this.netName = netName;
+
+    this.accountAddressArray = lodash.range(addressCount)
+      .map(i => ChecksumAddress.fromHex(accountAddressStruct.encode({ address: i }), netName).toString());
+    this.tokenAddressArray = lodash.range(epochTxCount)
+      .map(i => ChecksumAddress.fromHex(tokenAddressStruct.encode({ index: i }), netName).toString());
   }
 
   async call(method, ...params) {
@@ -135,18 +66,9 @@ class MockProvider extends EventEmitter {
     return 'mock';
   }
 
-  cfx_getSupplyInfo() {
-    return {
-      totalIssued: randomHex(20),
-      totalStaking: randomHex(18),
-      totalCollateral: randomHex(18),
-      totalCirculating: randomHex(18),
-    };
-  }
-
   cfx_getStatus() {
     return {
-      chainId: toHex(1),
+      chainId: toHex(this.chainId),
       epochNumber: randomHex(4),
       blockNumber: randomHex(4),
       pendingTxNumber: randomHex(4),
@@ -158,23 +80,32 @@ class MockProvider extends EventEmitter {
     return randomHex(2);
   }
 
-  cfx_getInterestRate() {
+  cfx_getSupplyInfo(epochNumber) {
+    return {
+      totalIssued: randomHex(20),
+      totalStaking: randomHex(18),
+      totalCollateral: randomHex(18),
+      totalCirculating: randomHex(18),
+    };
+  }
+
+  cfx_getInterestRate(epochNumber) {
     return randomHex(10);
   }
 
-  cfx_getAccumulateInterestRate() {
+  cfx_getAccumulateInterestRate(epochNumber) {
     return randomHex(10);
   }
 
   // ------------------------------- address ----------------------------------
-  cfx_getAccount() {
+  cfx_getAccount(address, epochNumber) {
     return {
       accumulatedInterestReturn: randomHex(4),
       balance: randomHex(8),
       collateralForStorage: randomHex(4),
       nonce: randomHex(2),
       stakingBalance: randomHex(2),
-      admin: '0x0000000000000000000000000000000000000000',
+      admin: ChecksumAddress.fromHex('0x0000000000000000000000000000000000000000', this.netName),
       codeHash: randomHex(64),
     };
   }
@@ -183,8 +114,8 @@ class MockProvider extends EventEmitter {
     return Number(epochNumber) ? randomHex(8) : '0x0';
   }
 
-  cfx_getStakingBalance() {
-    return randomHex(10);
+  cfx_getStakingBalance(address, epochNumber) {
+    return Number(epochNumber) ? randomHex(8) : '0x0';
   }
 
   cfx_getNextNonce(address, epochNumber) {
@@ -192,12 +123,12 @@ class MockProvider extends EventEmitter {
       return toHex(Number.MAX_SAFE_INTEGER);
     }
 
-    const number = (Number(epochNumber) * this.epochBlockCount * this.blockTxCount) / this.addressCount;
+    const number = (Number(epochNumber) * this.epochBlockCount * this.blockTxCount) / this.accountAddressArray.length;
     return toHex(Math.floor(number));
   }
 
-  cfx_getAdmin() {
-    return randomHex(40);
+  cfx_getAdmin(address, epochNumber) {
+    return ChecksumAddress.fromHex('0x0000000000000000000000000000000000000000', this.netName);
   }
 
   cfx_getVoteList(address, epochNumber) {
@@ -216,17 +147,23 @@ class MockProvider extends EventEmitter {
   }
 
   // -------------------------------- epoch -----------------------------------
-  cfx_epochNumber() {
-    return toHex(Number.MAX_SAFE_INTEGER);
+  cfx_epochNumber(epochNumber) {
+    if (/^0x[0-9a-f]+$/.test(epochNumber)) {
+      return epochNumber;
+    }
+    return toHex(this.epochNumber);
+  }
+
+  cfx_getBlocksByEpoch(epochNumber) {
+    const blockHashArray = lodash.range(this.epochBlockCount).map(
+      i => blockHashStruct.encode({ epochNumber, blockIndex: i }),
+    );
+    return blockHashArray.reverse(); // pivot to last one
   }
 
   cfx_getBlockByEpochNumber(epochNumber, detail) {
     const blockHashArray = this.cfx_getBlocksByEpoch(epochNumber);
     return this.cfx_getBlockByHash(lodash.last(blockHashArray), detail);
-  }
-
-  cfx_getBlocksByEpoch(epochNumber) {
-    return mockBlockHashArray(this, epochNumber);
   }
 
   cfx_getBlockRewardInfo(epochNumber) {
@@ -245,34 +182,48 @@ class MockProvider extends EventEmitter {
 
   // -------------------------------- block -----------------------------------
   cfx_getBestBlockHash() {
-    return randomHex(64); // XXX
+    return blockHashStruct.encode({ epochNumber: 0, blockIndex: 0 });
   }
 
   cfx_getBlockByHash(blockHash, detail = false) {
-    const { epochNumber, parentHash, refereeHashes, miner, timestamp } = mockBlockByHash(this, blockHash);
+    const { epochNumber, blockIndex } = blockHashStruct.decode(blockHash);
+    const blockCount = epochNumber * this.epochBlockCount + blockIndex;
+    const timestamp = this.startTimestamp + (blockCount * this.blockDelta); // in secords
+    const miner = this.accountAddressArray[blockCount % this.accountAddressArray.length];
 
-    let transactions = mockTxHashArray(this, blockHash);
+    const parentHash = blockIndex === 0
+      ? blockHashStruct.encode({ epochNumber: epochNumber ? epochNumber - 1 : 0, blockIndex })
+      : blockHashStruct.encode({ epochNumber, blockIndex: blockIndex - 1 });
+
+    const refereeHashes = [blockHashStruct.encode({ epochNumber, blockIndex })];
+
+    let transactions = lodash.range(this.blockTxCount).map(
+      i => txHashStruct.encode({ epochNumber, blockIndex, transactionIndex: i }),
+    );
+
     if (detail) {
       transactions = transactions.map(transactionHash => this.cfx_getTransactionByHash(transactionHash));
     }
 
     return {
-      adaptive: randomPick(true, false),
+      adaptive: lodash.sample([true, false]),
       blame: randomHex(1),
+      custom: [[1]],
       deferredLogsBloomHash: randomHex(64),
       deferredReceiptsRoot: randomHex(64),
       deferredStateRoot: randomHex(64),
       difficulty: randomHex(8),
-      epochNumber,
-      gasLimit: randomHex(2),
+      epochNumber: toHex(epochNumber),
+      gasLimit: randomHex(4),
+      gasUsed: randomHex(2),
       hash: blockHash,
-      height: epochNumber,
+      height: toHex(epochNumber),
       miner,
       nonce: randomHex(16),
       parentHash,
+      powQuality: randomHex(4),
       refereeHashes,
       size: randomHex(4),
-      powQuality: randomHex(4),
       timestamp,
       transactions,
       transactionsRoot: randomHex(64),
@@ -285,93 +236,105 @@ class MockProvider extends EventEmitter {
       throw new Error('{"code":-32602,"message":"Error: pivot chain assumption failed"}');
     }
 
-    const block = this.cfx_getBlockByHash(blockHash, true);
-    block.stable = randomPick(true, false);
-    block.transactions.forEach(transaction => {
-      transaction.status = randomPick('0x0', '0x01');
-    });
-    return block;
+    return this.cfx_getBlockByHash(blockHash, true);
   }
 
   cfx_getConfirmationRiskByHash() {
     return randomHex(64);
   }
 
+  trace_block(blockHash) {
+    const { transactions } = this.cfx_getBlockByHash(blockHash, true);
+    return {
+      transactionTraces: transactions.map(({ from, to, value, gas, data }) => ({
+        traces: [{
+          type: 'call',
+          action: { callType: 'call', from, to, value, gas, input: data },
+        }],
+      })),
+    };
+  }
+
   // ----------------------------- transaction --------------------------------
   cfx_getTransactionByHash(transactionHash) {
-    const { epochNumber, blockHash, from, nonce, to, transactionIndex, status } = mockTxByHash(this, transactionHash);
+    const { epochNumber, blockIndex, transactionIndex } = txHashStruct.decode(transactionHash);
+    const blockCount = epochNumber * this.epochBlockCount + blockIndex;
+    const txCount = blockCount * this.blockTxCount + transactionIndex;
+
+    const blockHash = blockHashStruct.encode({ epochNumber, blockIndex });
+    const nonce = Math.floor(txCount / this.accountAddressArray.length);
+    const from = this.accountAddressArray[txCount % this.accountAddressArray.length];
+    const to = transactionIndex === 0 ? null : this.accountAddressArray[(txCount + 1) % this.accountAddressArray.length];
+    const contractCreated = to ? null : ChecksumAddress.fromHex(contractAddressStruct.encode({ epochNumber, blockIndex }), this.netName).toString();
 
     return {
       blockHash,
-      contractCreated: null,
+      chainId: toHex(this.chainId),
+      contractCreated,
       data: randomHex(100),
+      epochHeight: epochNumber,
       from,
       gas: randomHex(5), // gasLimit
       gasPrice: randomHex(2),
       hash: transactionHash,
-      nonce,
+      nonce: toHex(nonce),
       r: randomHex(64),
       s: randomHex(64),
-      status,
-      to,
-      transactionIndex,
-      v: randomPick('0x0', '0x1'),
-      value: randomHex(4),
+      status: (blockIndex === 0 && transactionIndex === 1) ? '0x1' : '0x0',
       storageLimit: randomHex(10),
-      chainId: '0x0',
-      epochHeight: epochNumber,
+      to,
+      transactionIndex: toHex(transactionIndex),
+      v: lodash.sample(['0x0', '0x1']),
+      value: randomHex(4),
     };
   }
 
   cfx_getTransactionReceipt(transactionHash) {
-    const { blockHash, from, to, transactionIndex, status } = mockTxByHash(this, transactionHash);
-    const { epochNumber } = mockBlockByHash(this, transactionHash);
+    const { epochNumber } = txHashStruct.decode(transactionHash);
+    const { blockHash, contractCreated, from, to, transactionIndex, status } = this.cfx_getTransactionByHash(transactionHash);
 
     return {
       blockHash,
-      contractCreated: randomPick(null, randomHex(40)),
+      contractCreated,
       epochNumber: toHex(epochNumber),
       from,
-      gasUsed: randomHex(5),
+      gasCoveredBySponsor: lodash.sample([true, false]),
       gasFee: randomHex(5),
-      index: toHex(transactionIndex),
+      gasUsed: randomHex(5),
+      index: transactionIndex,
       logs: [],
       logsBloom: randomHex(512),
-      outcomeStatus: status === null ? null : toHex(status),
+      outcomeStatus: status,
       stateRoot: randomHex(64),
+      storageCollateralized: randomHex(4),
+      storageCoveredBySponsor: lodash.sample([true, false]),
+      storageReleased: [
+        { address: from, collaterals: randomHex(4) },
+      ],
       to,
       transactionHash,
       txExecErrorMsg: status === null ? null : 'mock error message',
-      gasCoveredBySponsor: false,
-      storageCoveredBySponsor: true,
-      storageCollateralized: randomHex(4),
-      storageReleased: [
-        {
-          address: from,
-          collaterals: randomHex(4),
-        },
-      ],
     };
   }
 
-  cfx_sendRawTransaction() {
+  cfx_sendRawTransaction(hex) {
     return randomHex(64);
   }
 
-  cfx_sendTransaction() {
+  cfx_sendTransaction(options) {
     return randomHex(64);
   }
 
   // ------------------------------ contract ----------------------------------
-  cfx_getCode() {
+  cfx_getCode(address, epochNumber) {
     return randomHex(100);
   }
 
-  cfx_getStorageAt() {
+  cfx_getStorageAt(address, position, epochNumber) {
     return randomHex(64);
   }
 
-  cfx_getStorageRoot() {
+  cfx_getStorageRoot(address, epochNumber) {
     return {
       delta: randomHex(64),
       intermediate: randomHex(64),
@@ -379,25 +342,25 @@ class MockProvider extends EventEmitter {
     };
   }
 
-  cfx_getSponsorInfo() {
+  cfx_getSponsorInfo(address, epochNumber) {
     return {
       sponsorBalanceForCollateral: randomHex(2),
       sponsorBalanceForGas: randomHex(2),
       sponsorGasBound: randomHex(2),
-      sponsorForCollateral: '0x0000000000000000000000000000000000000000',
-      sponsorForGas: '0x0000000000000000000000000000000000000000',
+      sponsorForCollateral: ChecksumAddress.fromHex('0x0000000000000000000000000000000000000000', this.netName).toString(),
+      sponsorForGas: ChecksumAddress.fromHex('0x0000000000000000000000000000000000000000', this.netName).toString(),
     };
   }
 
-  cfx_getCollateralForStorage() {
+  cfx_getCollateralForStorage(address, epochNumber) {
     return randomHex(10);
   }
 
-  cfx_call() {
+  cfx_call(hex) {
     return padHex(100, 64);
   }
 
-  cfx_estimateGasAndCollateral() {
+  cfx_estimateGasAndCollateral(hex, epochNumber) {
     return {
       gasUsed: randomHex(4),
       gasLimit: randomHex(4),
@@ -405,23 +368,41 @@ class MockProvider extends EventEmitter {
     };
   }
 
-  cfx_getLogs({ fromEpoch, topics }) {
+  cfx_getLogs({ blockHashes, address, topics, fromEpoch, toEpoch, limit }) {
     const epochNumber = Number(fromEpoch); // Number or NaN
-
-    if (Number.isInteger(epochNumber)) {
-      return mockLogsByEpochNumber(this, epochNumber, topics);
+    if (!Number.isInteger(epochNumber)) {
+      return [];
     }
 
-    return [];
+    const { hash: blockHash, transactions: [transactionHash] } = this.cfx_getBlockByEpochNumber(epochNumber);
+
+    return lodash.range(this.epochTxCount).map(index => {
+      const { topics: topicArray, data } = contract.Transfer(
+        accountAddressStruct.encode({ address: index }),
+        accountAddressStruct.encode({ address: index + 1 }),
+        epochNumber,
+      );
+
+      return {
+        address: this.tokenAddressArray[index],
+        blockHash,
+        data,
+        epochNumber: toHex(epochNumber),
+        logIndex: randomHex(1),
+        topics: topicArray,
+        transactionHash,
+        transactionIndex: toHex(0),
+        transactionLogIndex: toHex(index),
+      };
+    });
   }
 
-  // ------------------------------ subscribe ---------------------------------
   cfx_subscribe() {
     return randomHex(16);
   }
 
   cfx_unsubscribe() {
-    return randomPick(true, false);
+    return lodash.sample([true, false]);
   }
 }